+++ index.js
@@ -17,6 +17,7 @@
  * limitations under the License.
  */
 
+export { AppData } from './src/Application'
 export { default as Application } from './src/Application'
 export { default as AudioPlayer } from './src/AudioPlayer'
 export { default as Img } from './src/Img'
@@ -27,8 +28,11 @@ export { default as Locale } from './src/Locale'
 export { default as Log } from './src/Log'
 export { default as MediaPlayer } from './src/MediaPlayer'
 export { default as Metrics } from './src/Metrics'
+export { default as Pin } from './src/Pin'
 export { default as Profile } from './src/Profile'
 export { default as Purchase } from './src/Purchase'
+export { default as Router } from './src/Router'
 export { default as Settings } from './src/Settings'
 export { default as Storage } from './src/Storage'
+export { default as TV } from './src/TV'
 export { default as Utils } from './src/Utils'