+++ src/Application/index.js
@@ -24,6 +24,12 @@ import Metrics from '../Metrics'
 import VersionLabel from '../VersionLabel'
 import FpsCounter from '../FpsCounter'
 import Log from '../Log'
+import Settings from '../Settings'
+
+import { version as sdkVersion } from '../../package.json'
+
+export let AppInstance
+export let AppData
 
 const defaultOptions = {
   stage: { w: 1920, h: 1080, clearColor: 0x00000000, canvas2d: false },
@@ -76,18 +82,26 @@ export default function(App, appData, platformSettings) {
       ])
         .then(() => {
           Metrics.app.loaded()
-          this.childList.a({
+
+          AppData = appData
+
+          AppInstance = this.stage.c({
             ref: 'App',
             type: App,
-            appData,
             forceZIndexContext: !!platformSettings.showVersion || !!platformSettings.showFps,
           })
+
+          this.childList.a(AppInstance)
+
+          Log.info('App version', this.config.version)
+          Log.info('SDK version', sdkVersion)
 
           if (platformSettings.showVersion) {
             this.childList.a({
               ref: 'VersionLabel',
               type: VersionLabel,
               version: this.config.version,
+              sdkVersion: sdkVersion,
             })
           }
 
@@ -112,6 +126,10 @@ export default function(App, appData, platformSettings) {
     }
 
     closeApp() {
+      Log.info('Closing App')
+
+      Settings.clearSubscribers()
+
       if (platformSettings.onClose && typeof platformSettings.onClose === 'function') {
         platformSettings.onClose()
       } else {
@@ -144,8 +162,13 @@ export default function(App, appData, platformSettings) {
       })
     }
 
+    set focus(v) {
+      this._focussed = v
+      this._refocus()
+    }
+
     _getFocused() {
-      return this.tag('App')
+      return this._focussed || this.tag('App')
     }
   }
 }