+++ src/Launch/index.js
@@ -23,7 +23,11 @@ import { initMetrics } from '../Metrics'
 import { initSettings } from '../Settings'
 import { initMediaPlayer } from '../MediaPlayer'
 import { initStorage } from '../Storage'
+import { initRouter } from '../Router'
+import { initTV } from '../TV'
 import Application from '../Application'
+
+export let ApplicationInstance
 
 export default (App, appSettings, platformSettings, appData) => {
   initSettings(appSettings, platformSettings)
@@ -36,8 +40,11 @@ export default (App, appSettings, platformSettings, appData) => {
     platformSettings.plugins.profile && initProfile(platformSettings.plugins.profile)
     platformSettings.plugins.metrics && initMetrics(platformSettings.plugins.metrics)
     platformSettings.plugins.mediaPlayer && initMediaPlayer(platformSettings.plugins.mediaPlayer)
+    platformSettings.plugins.router && initRouter(platformSettings.plugins.router)
+    platformSettings.plugins.tv && initTV(platformSettings.plugins.tv)
   }
 
   const app = Application(App, appData, platformSettings)
-  return new app(appSettings)
+  ApplicationInstance = new app(appSettings)
+  return ApplicationInstance
 }