+++ support/startApp.js
@@ -28,7 +28,11 @@ var _this = undefined;
  */
 var style = document.createElement('style');
 document.head.appendChild(style);
-style.sheet.insertRule('@media all { html {height: 100%; width: 100%;} *,body {margin:0; padding:0;} canvas { position: absolute; z-index: 2; } body { background: black; width: 100%; height: 100%;} }');
+style.sheet.insertRule('@media all { html {height: 100%; width: 100%;} *,body {margin:0; padding:0;} canvas { position: absolute; z-index: 2; } body { width: 100%; height: 100%;} }');
+var app;
+var canvas;
+var appMetadata;
+var settings;
 
 var startApp = function startApp() {
   var _this2 = this;
@@ -36,8 +40,6 @@ var startApp = function startApp() {
   _newArrowCheck(this, _this);
 
   console.time('app');
-  var appMetadata;
-  var settings;
   sequence([function () {
     var _this3 = this;
 
@@ -79,7 +81,7 @@ var startApp = function startApp() {
 
     _newArrowCheck(this, _this2);
 
-    return hasTextureMode().then(function (enabled) {
+    return hasTextureMode(settings.platformSettings).then(function (enabled) {
       _newArrowCheck(this, _this5);
 
       return settings.platformSettings.textureMode = enabled;
@@ -100,8 +102,9 @@ var startApp = function startApp() {
     console.time('app2');
     settings.appSettings.version = appMetadata.version;
     settings.appSettings.id = appMetadata.identifier;
-    var app = window[appMetadata.id](settings.appSettings, settings.platformSettings, settings.appData);
-    document.body.appendChild(app.stage.getCanvas());
+    app = window[appMetadata.id](settings.appSettings, settings.platformSettings, settings.appData);
+    canvas = app.stage.getCanvas();
+    document.body.appendChild(canvas);
   }.bind(this)]);
 }.bind(undefined);
 
@@ -110,11 +113,7 @@ var getAppMetadata = function getAppMetadata() {
 
   _newArrowCheck(this, _this);
 
-  return fetch('./metadata.json').then(function (response) {
-    _newArrowCheck(this, _this7);
-
-    return response.json();
-  }.bind(this)).then(function (metadata) {
+  return fetchJson('./metadata.json').then(function (metadata) {
     _newArrowCheck(this, _this7);
 
     metadata.id = "APP_".concat(metadata.identifier.replace(/[^0-9a-zA-Z_$]/g, '_'));
@@ -127,21 +126,84 @@ var getSettings = function getSettings() {
 
   _newArrowCheck(this, _this);
 
-  return fetch('./settings.json').then(function (response) {
+  return new Promise(function (resolve) {
+    var _this9 = this;
+
     _newArrowCheck(this, _this8);
 
-    return response.json();
-  }.bind(this)).catch(function (error) {
-    _newArrowCheck(this, _this8);
+    var settings;
+    fetchJson('./settings.json').then(function (json) {
+      _newArrowCheck(this, _this9);
 
-    console.warn('No settings.json found. Using defaults.');
-    return {
-      appSettings: {},
-      platformSettings: {
-        path: './static',
-        esEnv: 'es6'
-      }
-    };
+      return settings = json;
+    }.bind(this)).catch(function () {
+      _newArrowCheck(this, _this9);
+
+      console.warn('No settings.json found. Using defaults.');
+      settings = {
+        appSettings: {},
+        platformSettings: {
+          path: './static',
+          esEnv: 'es6'
+        }
+      };
+    }.bind(this)).finally(function () {
+      var _this10 = this;
+
+      _newArrowCheck(this, _this9);
+
+      settings.platformSettings = settings.platformSettings || {};
+
+      settings.platformSettings.onClose = function () {
+        var _this11 = this;
+
+        _newArrowCheck(this, _this10);
+
+        // clean up video
+        var videoElements = document.getElementsByTagName('video');
+
+        if (videoElements.length) {
+          videoElements[0].src = '';
+        } // signal to close app
+
+
+        app.close && app.close(); // clear canvas
+
+        if (canvas) {
+          var stage = app.stage; // maybe move this to a plugin so we can customize it per platform
+
+          if (stage.gl) {
+            stage.gl.clearColor(0.0, 0.0, 0.0, 0.0);
+            stage.gl.clear(stage.gl.COLOR_BUFFER_BIT);
+          } else {
+            stage.c2d.clearRect(0, 0, canvas.width, canvas.height);
+          }
+        } // cleanup
+
+
+        setTimeout(function () {
+          _newArrowCheck(this, _this11);
+
+          if (canvas) {
+            canvas.remove();
+          } // detach app bundle from window scope
+
+
+          window[appMetadata.id] = null; // remove script tag
+
+          removeJS('appbundle'); // reset vars
+
+          app = null;
+          canvas = null;
+          appMetadata = null;
+          settings = null; // show notice to refresh
+
+          console.log('👋 App closed!\nRefresh the page to restart the App');
+        }.bind(this));
+      }.bind(this);
+
+      resolve(settings);
+    }.bind(this));
   }.bind(this));
 }.bind(undefined); // FIXME: these 3 functions could be refactored to a single one receiving 2 arguments (filename, esEnv)
 
@@ -157,7 +219,7 @@ var loadAppBundle = function loadAppBundle(esEnv) {
   _newArrowCheck(this, _this);
 
   var filename = !esEnv || esEnv === 'es6' ? './appBundle.js' : './appBundle.' + esEnv + '.js';
-  return loadJS(filename);
+  return loadJS(filename, 'appbundle');
 }.bind(undefined);
 
 var loadLightningInspect = function loadLightningInspect(esEnv) {
@@ -168,20 +230,24 @@ var loadLightningInspect = function loadLightningInspect(esEnv) {
 }.bind(undefined);
 
 var loadPolyfills = function loadPolyfills(esEnv) {
-  var _this9 = this;
+  var _this12 = this;
 
   _newArrowCheck(this, _this);
 
   // load polyfills when esEnv is defined and it's not es6
   if (esEnv && esEnv !== 'es6') {
     return sequence([function () {
-      _newArrowCheck(this, _this9);
+      _newArrowCheck(this, _this12);
 
       return loadJS('./polyfills/babel-polyfill.js');
     }.bind(this), function () {
-      _newArrowCheck(this, _this9);
+      _newArrowCheck(this, _this12);
 
       return loadJS('./polyfills/url.js');
+    }.bind(this), function () {
+      _newArrowCheck(this, _this12);
+
+      return loadJS('./polyfills/fetch.js');
     }.bind(this)]);
   }
 
@@ -189,12 +255,12 @@ var loadPolyfills = function loadPolyfills(esEnv) {
 }.bind(undefined);
 
 var loadJS = function loadJS(url, id) {
-  var _this10 = this;
+  var _this13 = this;
 
   _newArrowCheck(this, _this);
 
   return new Promise(function (resolve) {
-    _newArrowCheck(this, _this10);
+    _newArrowCheck(this, _this13);
 
     console.log('loadJS', url);
     var tag = document.createElement('script');
@@ -205,33 +271,65 @@ var loadJS = function loadJS(url, id) {
   }.bind(this));
 }.bind(undefined);
 
+var removeJS = function removeJS(id) {
+  _newArrowCheck(this, _this);
+
+  var scriptEl = document.getElementById(id);
+
+  if (scriptEl) {
+    scriptEl.remove();
+  }
+}.bind(undefined);
+
+var fetchJson = function fetchJson(file) {
+  var _this14 = this;
+
+  _newArrowCheck(this, _this);
+
+  return new Promise(function (resolve, reject) {
+    _newArrowCheck(this, _this14);
+
+    var xhr = new XMLHttpRequest();
+
+    xhr.onreadystatechange = function () {
+      if (xhr.readyState == XMLHttpRequest.DONE) {
+        if (xhr.status === 200) resolve(JSON.parse(xhr.responseText));else reject(xhr.statusText);
+      }
+    };
+
+    xhr.open('GET', file);
+    xhr.send(null);
+  }.bind(this));
+}.bind(undefined);
+
 var sequence = function sequence(steps) {
-  var _this11 = this;
+  var _this15 = this;
 
   _newArrowCheck(this, _this);
 
   return steps.reduce(function (promise, method) {
-    var _this12 = this;
+    var _this16 = this;
 
-    _newArrowCheck(this, _this11);
+    _newArrowCheck(this, _this15);
 
     return promise.then(function () {
-      _newArrowCheck(this, _this12);
+      _newArrowCheck(this, _this16);
 
       return method();
     }.bind(this));
   }.bind(this), Promise.resolve(null));
 }.bind(undefined);
 
-var hasTextureMode = function hasTextureMode() {
-  var _this13 = this;
+var hasTextureMode = function hasTextureMode(platformSettings) {
+  var _this17 = this;
 
   _newArrowCheck(this, _this);
 
   return new Promise(function (resolve) {
-    _newArrowCheck(this, _this13);
+    _newArrowCheck(this, _this17);
 
-    // yes, this could be a oneliner, but zebra es5 couldn't handle that (so 2 lines to be safe)
+    if (platformSettings.textureMode === true) resolve(true); // yes, this could be a oneliner, but zebra es5 couldn't handle that (so 2 lines to be safe)
+
     var url = new URL(document.location.href);
     resolve(url.searchParams.has('texture'));
   }.bind(this));